[SerialPort] moved ‘close()’ to operation
I didn’t want delegates to have to manage closing the port themselves. It seemed weird to have them remember to do that, or care at all about it.

Also, removed the ability to specify the ‘timeout’ amount; again, the controller should know its hardware best, and the caller shouldn’t make assumptions about the bus timing of the controller.