[SerialPort] ultimate ‘close()’ fix! 🎉
This solution works by leveraging the serial port delegate (which I’ve should’ve been doing all along). It will wait until the port is closed, then send the result, as well as its own delegate callback, to any of the interested parties.

This fixes any potential “hangs” that could occur previously, when an operation was waiting to start, but received the ‘portClosed’ callback of the previous operation.