[CartridgeController] purge ‘Platform’ association
No more `associatedtype Platform` for `CartridgeController`. This means that `insideGadgetsController` is no longer generic.

This is a pretty massive change, since now the APIs return `Result` types, rather than relying on callbacks (most which used to be parameterized as `result`). More importantly, the `perform` function helps eliminate the many `DispatchQueue/Operation` “hops” that were occurring. `perform` is always called on its own `DispatchQueue`. As a result, this change prevents a caller from using the `try await { }`, which is risky and could deadlock the main thread if the user wasn’t aware of the consequences.